requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.23.2
pydantic==2.11.7
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
motor==3.3.2
pymongo==4.6.1
pdfplumber==0.9.0
python-docx==0.8.11
httpx==0.24.1
